# api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUserRequest:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  models.DeleteUserResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.GetUserResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.GetUsersResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      message:
        type: string
      refresh_token:
        type: string
      success:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.Message:
    properties:
      err:
        type: string
    type: object
  models.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  models.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      message:
        type: string
      refresh_token:
        type: string
      success:
        type: boolean
    type: object
...
  models.UpdateUserRequest:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserFilter:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: Api-gateway service
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: This method logs in a user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: LOGIN USER
      tags:
      - AUTH
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: This method refreshes the authentication token
      parameters:
      - description: Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: REFRESH TOKEN
      tags:
      - AUTH
  /auth/register:
    post:
      consumes:
      - application/json
      description: This method registers a new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: REGISTER USER
      tags:
      - AUTH
...
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: This method deletes a user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: DELETE USER
      tags:
      - USER
  /users:
    get:
      consumes:
      - application/json
      description: This method gets all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET USERS
      tags:
      - USER
  /users/{id}:
    get:
      consumes:
      - application/json
      description: This method gets a user by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET USER
      tags:
      - USER
    put:
      consumes:
      - application/json
      description: This method updates a user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: UPDATE USER
      tags:
      - USER
  /users/filter:
    post:
      consumes:
      - application/json
      description: This method gets users by filter
      parameters:
      - description: Filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/models.UserFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET USERS BY FILTER
      tags:
      - USER
securityDefinitions:
  BearerAuth:
    in: header
    name: Authourization
    type: apiKey
swagger: "2.0"